+++ languages/en/tl_submission.php
@@ -36,6 +36,7 @@ $arrLang['stopDatime']              = ['End date', 'Select a end date.'];
 $arrLang['addDifferentBillingData'] = ['Specify different billing data', 'Choose this option to specify different billing data.'];
 $arrLang['published']               = ['Activated', 'Choose this option to activate the submission.'];
 $arrLang['formHybridBlob']          = ['FormHybrid BLOB', 'This field contains changes, that are not supposed to be finally saved into the record (e.g. multistep forms).'];
+$arrLang['submissionLanguage']      = ['Language of submission', 'This field contains the language in which the submission was edited.'];
 
 /**
  * Legends