+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.7.0] - 2020-03-19
+- added oncreate_callback in `SubmissionModel::create` and `tl_submission`
+- added `submissionLanguage` to `tl_submission`
+
 ## [1.6.2] - 2020-02-14
 - fixed `attachmentExtensions` sql